streamlit-debate-app/requirements.txt
streamlit
openai
requests
python-dotenv
pandas
numpy